titan-poss-poss-web-f78a2c9ef984/titan-poss-poss-web-f78a2c9ef984/libs/shared/order-confirmation/data-access-order-confirmation/src/lib/+state/order-confirmation.selectors.ts
import { createSelector } from '@ngrx/store';
import { selectOrderConfirmationState } from './order-confirmation.reducer';

const selectHasError = createSelector(
  selectOrderConfirmationState,
  state => state.hasError
);

const selectIsLoading = createSelector(
  selectOrderConfirmationState,
  state => state.isLoading
);

const updateCashMemoResponse = createSelector(
  selectOrderConfirmationState,
  state => state.updateCashMemoResponse
);



export const orderConfirmationSelectors = {
  selectHasError,
  selectIsLoading,
  updateCashMemoResponse

};
